requirements-dev.txt
Django==1.11.2
# graphene==1.4.1
https://github.com/graphql-python/graphene/archive/master.zip

graphene-django==1.3
# For OAuth token login
django-oauth-toolkit